Moves FigmaBundle to a separate file
