+++ FigmaSharp/FigmaSharp/FigmaManifest.cs
@@ -3,7 +3,6 @@ using System.Text;
 using System.Linq;
 using Newtonsoft.Json;
 using System.IO;
-using FigmaSharp.Models;
 using System.Threading.Tasks;
 
 namespace FigmaSharp
@@ -15,137 +14,6 @@ namespace FigmaSharp
 		public ManifestDescription (string description)
 		{
 			this.Description = description;
-		}
-	}
-
-	public class FigmaBundle
-	{
-		public FigmaManifest Manifest { get; set; }
-
-		public string DirectoryPath { get; private set; }
-
-		public string Name => DirectoryPath == null ? null : Path.GetFileName (DirectoryPath);
-
-		internal const string FigmaBundleDirectoryExtension = ".figmabundle";
-		//const string FigmaBundlesDirectoryName = ".bundles";
-		internal const string FigmaDirectoryName = ".figma";
-		internal const string ManifestFileName = "manifest.json";
-		internal const string DocumentFileName = "document.figma";
-		internal const string ResourcesDirectoryName = "Resources";
-		internal const string ImageFormat = ".png";
-
-		public void Load (string bundleDirectoryPath)
-		{
-			if (!Directory.Exists (bundleDirectoryPath)) {
-				throw new DirectoryNotFoundException ("directory doesn't exists");
-			}
-
-			var manifestFullPath = Path.Combine (bundleDirectoryPath, ManifestFileName);
-			if (!File.Exists (manifestFullPath)) {
-				throw new FileNotFoundException ("manifest doesn't exists");
-			}
-
-			DirectoryPath = bundleDirectoryPath;
-
-			try {
-				Manifest = FigmaManifest.FromFilePath (manifestFullPath);
-			} catch (Exception ex) {
-				throw new FileLoadException ("error reading manifest file", ex);
-			}
-		}
-
-		public void Save ()
-		{
-			if (DirectoryPath == null) {
-				return;
-			}
-
-			if (!Directory.Exists (DirectoryPath)) {
-				Directory.CreateDirectory (DirectoryPath);
-			}
-
-			if (Manifest != null) {
-				Manifest.Save (Path.Combine (DirectoryPath, ManifestFileName));
-			}
-		}
-
-		public static FigmaBundle Create (string fileId, string directoryPath)
-		{
-			var bundle = new FigmaBundle () {
-				DirectoryPath = directoryPath
-			};
-			bundle.Manifest = new FigmaManifest () {
-				ApiVersion = AppContext.Current.Version,
-				Date = DateTime.Now,
-				DocumentUrl = fileId
-			};
-
-			return bundle;
-		}
-
-		public static FigmaBundle FromDirectoryPath (string fullPath)
-		{
-			try {
-				var bundle = new FigmaBundle ();
-				bundle.Load (fullPath);
-				return bundle;
-			} catch (Exception ex) {
-				Console.WriteLine (ex);
-				//not a bundle
-				return null;
-			}
-		}
-
-		public string FileId => Manifest.DocumentUrl;
-
-		public string DocumentFilePath => Path.Combine (DirectoryPath, DocumentFileName);
-
-		internal static FigmaResponse GenerateDocumentOutputFile (string fileId, string directoryPath)
-		{
-			var documentFilePath = Path.Combine (directoryPath, DocumentFileName);
-			if (File.Exists (documentFilePath))
-				File.Delete (documentFilePath);
-			//generate also Document.figma
-			var content = FigmaApiHelper.GetFigmaFileContent (fileId);
-		
-			File.WriteAllText (documentFilePath, content);
-			var figmaResponse = FigmaApiHelper.GetFigmaResponseFromContent (content);
-			return figmaResponse;
-		}
-
-		internal static void GenerateOutputResourceFiles (string fileId, FigmaResponse figmaResponse, string resourcesDirectoryPath)
-		{
-			var mainNode = figmaResponse.document.children.FirstOrDefault ();
-
-			var figmaImageIds = mainNode.OfTypeImage ()
-				.Select (s => s.id)
-				.ToArray ();
-
-			if (figmaImageIds.Length > 0) {
-				
-				if (!Directory.Exists (resourcesDirectoryPath)) {
-					Directory.CreateDirectory (resourcesDirectoryPath);
-				}
-
-				var figmaImageResponse = FigmaApiHelper.GetFigmaImages (fileId, figmaImageIds);
-				FileHelper.SaveFiles (resourcesDirectoryPath, ImageFormat, figmaImageResponse.images);
-			}
-		}
-
-		internal void GenerateLocalDocument (bool includeImages)
-		{
-			if (string.IsNullOrEmpty (FileId)) {
-				throw new InvalidOperationException ("id not set");
-			}
-
-			//get image resources
-			var figmaResponse = GenerateDocumentOutputFile (Manifest.DocumentUrl, DirectoryPath);
-
-			if (!includeImages)
-				return;
-
-			var resourcesDirectoryPath = Path.Combine (DirectoryPath, ResourcesDirectoryName);
-			GenerateOutputResourceFiles (FileId, figmaResponse, resourcesDirectoryPath);
 		}
 	}
